+++ src/js/library.js
@@ -2,7 +2,7 @@ import './refs';
 import { getWatchedItems } from './library-buttons';
 import { createFilmsGallery } from './markups';
 import pagination from './pagination';
-import { first, deletePageButton } from './pagination-layout';
+import { deletePageButton } from './pagination-layout';
 import './plug-for-library';
 import { createModalFilm } from './markups';
 import './modal';
@@ -31,33 +31,32 @@ pagination.on('beforeMove', event => {
   gallery.innerHTML = markup;
 });
 
-pagination.on('afterMove', event => {
-  topFunction();
-  // получаем номер активной страницы на кнопках
-  const currentPage = event.page;
+// pagination.on('afterMove', event => {
+//   topFunction();
+//   // получаем номер активной страницы на кнопках
+//   const currentPage = event.page;
 
-  // создаем массив для рендера по 20 айтемов на страницу
-  const arrayForMarkup = paginate(getWatchedItems(), 20, currentPage);
+//   // создаем массив для рендера по 20 айтемов на страницу
+//   const arrayForMarkup = paginate(getWatchedItems(), 20, currentPage);
 
-  totalPages = Math.ceil(arrayForMarkup.length / 20);
+//   totalPages = Math.ceil(arrayForMarkup.length / 20);
 
-  const first = document.querySelector('.tui-ico-first');
-  if (null !== first) {
-    first.textContent = '1';
-  }
+//   const first = document.querySelector('.tui-ico-first');
+//   if (null !== first) {
+//     first.textContent = '1';
+//   }
 
-  const last = document.querySelector('.tui-pagination .tui-ico-last');
-  if (null !== last) {
-    last.textContent = totalPages;
-  }
+//   const last = document.querySelector('.tui-pagination .tui-ico-last');
+//   if (null !== last) {
+//     last.textContent = totalPages;
+//   }
 
-  deletePageButton(1, '.tui-first');
-  deletePageButton(totalPages, '.tui-last');
+//   deletePageButton(1, '.tui-first');
+//   deletePageButton(totalPages, '.tui-last');
 
-  // const markup = createFilmsGallery(arrayForMarkup);
-  // gallery.innerHTML = markup;
-  renderPagin();
-});
+//   // const markup = createFilmsGallery(arrayForMarkup);
+//   // gallery.innerHTML = markup;
+// });
 
 // function renderPagin() {
 //   first();
